sketching out mvc routes
