+++ package.json
@@ -1,20 +1,20 @@
 {
-  "name": "my-app",
-  "version": "1.0.0",
-  "description": "",
-  "main": "server.js",
-  "dependencies": {
-    "concurrently": "^5.2.0",
-    "express": "^4.17.1"
-  },
-  "devDependencies": {},
-  "scripts": {
-    "client-install": "cd client && npm install",
-    "start": "node server.js",
-    "server": "nodemon server.js",
-    "client": "npm start --prefix client",
-    "dev": "concurrently \"npm run server\" \"npm run client\""
-  },
-  "author": "",
-  "license": "ISC"
+	"name": "my-app",
+	"version": "1.0.0",
+	"description": "",
+	"main": "server.js",
+	"dependencies": {
+		"concurrently": "^5.2.0",
+		"express": "^4.17.1"
+	},
+	"devDependencies": {},
+	"scripts": {
+		"client-install": "cd client && npm install",
+		"start": "node server.js",
+		"server": "cd server && nodemon index.js",
+		"client": "npm start --prefix client",
+		"dev": "concurrently \"npm run server\" \"npm run client\""
+	},
+	"author": "",
+	"license": "ISC"
 }
